requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.0.0
django-cors-headers==4.4.0
django-rest-knox==5.0.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
gunicorn==22.0.0
packaging==24.1
djangorestframework-simplejwt